requirements.txt
pyqt5>=5.15.0
python-dotenv>=0.19.0
requests>=2.27.0
pillow>=9.0.0
openai>=1.1.0
python-twitter>=3.5.0
schedule>=1.1.0
